Split gene and snp tests from main genomic test file.
